[FE] fix: 버그 삭제
