+++ FE/src/api/fetchWetboonInfo.ts
@@ -1,7 +1,7 @@
 import { WebtoonConfig, WebtoonResponseConfig } from "@/interface/Webtoon.interface";
 import axios from "axios";
 
-const fetchWetboonInfo = async (title: string): Promise<WebtoonConfig> => {
+const fetchWetboonInfo = async (title: string): Promise<WebtoonConfig[]> => {
   const response = await axios.get(`https://korea-webtoon-api.herokuapp.com/search?keyword=${title}`);
   const webttonData = response.data.webtoons || [];
   const filteredWebtoons = webttonData.map((webtoon: WebtoonResponseConfig) => ({
@@ -11,7 +11,7 @@ const fetchWetboonInfo = async (title: string): Promise<WebtoonConfig> => {
     updateDays: webtoon.updateDays,
     fanCount: webtoon.fanCount,
   }));
-  return filteredWebtoons;
+  return [filteredWebtoons];
 };
 
 export default fetchWetboonInfo;
