+++ FE/src/pages/mypage/EditLikedWtnPage.tsx
@@ -1,7 +1,8 @@
 import Text from "@/components/common/Text";
 import styles from "@/styles/signup/Signup.module.scss";
 import { useNavigate } from "react-router-dom";
-import React, {  useState } from "react";
+import React, { useEffect, useState } from "react";
+import fetchWetboonInfo from "@/api/fetchWetboonInfo";
 import SelectedWebtoonBox from "@/components/Webtoon/SelectedWebtoonBox";
 import SearchWebtoonContainer from "@/components/Webtoon/SearchWebtoonBox";
 import { WebtoonConfig } from "@/interface/Webtoon.interface";
@@ -18,19 +19,18 @@ const EditLikedWtnPage = () => {
     setSearch(e.target.value);
   };
 
-  // useEffect(() => {
-  //   // const fetchWebtoons = async () => {
-  //   //   try {
-  //   //     const response = await fetchWetboonInfo(search);
-  //   //     setWebtoons(response);
-  //   //   } catch (e) {
-  //   //     console.error("오류 발생", e);
-  //   //   }
-  //   // };
-  //   setWebtoons(response);
-  //   if (search) fetchWebtoons();
-  //   else setWebtoons([]);
-  // }, [search]);
+  useEffect(() => {
+    const fetchWebtoons = async () => {
+      try {
+        const response = await fetchWetboonInfo(search);
+        setWebtoons(response);
+      } catch (e) {
+        console.error("오류 발생", e);
+      }
+    };
+    if (search) fetchWebtoons();
+    else setWebtoons([]);
+  }, [search]);
 
   const handleSelect = (webtoon: WebtoonConfig) => {
     if (user.likedWebToonList.length >= 4) {
